Subscribe to a Platform Event in an Apex Trigger
Create an Apex trigger named OrderEventTrigger for Order_Event__e. This trigger will be similar to the CloudNewsTrigger trigger, but operates on the Order_Event__e event and creates tasks instead of cases.
If the order has shipped (event.Has_Shipped__c == true), create a task with the following values:
   -Priority: 'Medium'
   -Subject: 'Follow up on shipped order ' + event.Order_Number__c
   -OwnerId: event.CreatedById
   -(Note: You are assigning the task OwnerId to the same user who published the event. This step keep things simple so you don't have to perform any prerequisites.)