Add a Mobile Card that Displays Content Based on Opportunity Stage
To help the sales team with common objections at each stage of the sales process, create a mobile card with scripts and helpful content. Create a Visualforce page which will display a different piece of content depending on the current stage of the opportunity and add it as a mobile card to the Opportunity layout. The Visualforce page:

- The Visualforce page must be called 'OpportunityStageTips'.
- 'OpportunityStageTips' must evaluate the current stage of the opportunity and display different tips for the Prospecting, Needs Analysis, Proposal/Price Quote and Negotiation/Review stages. You can use any placeholder content to display for each stage.
- 'OpportunityStageTips' must use the Opportunity standard controller.
- 'OpportunityStageTips' must be added as a mobile card to the 'Opportunity Layout' page layout.